Update groups landing page
Formatted the top-level groups with h2 rather than having a long page of nested lists. This helps to make the structure of the information visually clearer (in the nested lists all the links are the same font size)
- Removed the TF summary sentences in the nested lists to reduce information overload on the landing page. The TF objectives are available right at the top of each page, when people click through.
This also helps the 'WAI-IG' and 'Previous groups' pages stand out and not be buried by all the TF information on this higher level landing page.